+++ tests/fixtures/fetch.php
@@ -1,1 +1,2 @@
+<?php use function Glowy\View\e ?>
 <?= e($this->fetch('foo')) ?>