+++ tests/fixtures/fetch_first.php
@@ -1,1 +1,2 @@
+<?php use function Glowy\View\e ?>
 <?= e($this->fetchFirst(['bar', 'foo'])) ?>