+++ tests/fixtures/fetch_unless.php
@@ -1,1 +1,2 @@
+<?php use function Glowy\View\e ?>
 <?= e($this->fetchUnless(false, 'foo')) ?>