+++ tests/fixtures/fetch_with_data.php
@@ -1,0 +1,3 @@
+<?php use function Glowy\View\e ?>
+<?= e($this->fetch('foo')) ?>
+<?= e($foo) ?>