+++ tests/fixtures/macro.php
@@ -1,1 +1,2 @@
+<?php use function Glowy\View\e ?>
 <?= e($foo) ?> <?= e($bar) ?>