+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+<a name="2.0.0"></a>
+# [2.0.0](https://github.com/glowyphp/view) (2022-07-04)
+* All Helpers functions are placed into the Glowy/View namespace.
+* Use union types.
+* Moving to PHP 8.1
+* Added new public function `file` to set view file.
+* Added new public function `data` to set view data.
+
 <a name="1.0.0"></a>
 # [1.0.0](https://github.com/glowyphp/view) (2022-02-05)
 * Initial release